Code/requirements.txt
flask
python-telegram-bot
tensorflow==2.1.0
numpy==1.18.1
pandas==1.0.1
protobuf==3.11.3
